+++ frontend/src/components/authentication/authComponents.tsx
@@ -1,35 +1,5 @@
 import React, { ReactChild } from 'react';
 import styled from 'styled-components';
-import { BiArrowBack } from 'react-icons/bi';
-
-const StyledBack = styled(BiArrowBack)`
-  margin-top: 10px;
-  color: black;
-`;
-
-const AuthHeader = styled.h1`
-  font-family: Poppins;
-  font-weight: 500;
-  font-size: 27px;
-  color: black;
-`;
-const Div = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  padding: 0;
-  @media only screen and (min-width: 768px) {
-    justify-content: center;
-    align-items: center;
-  }
-`;
-
-const AuthContent = styled.div`
-  padding: 0;
-  @media only screen and (min-width: 768px) {
-    padding: 40px;
-  }
-`;
 
 const StyledContainer = styled.div`
   text-align: left;
@@ -40,6 +10,17 @@ const StyledContainer = styled.div`
     background: #fffdfd;
     border: 1px solid #c4c4c4;
     border-radius: 7px;
+  }
+`;
+
+const Div = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  padding: 0;
+  @media only screen and (min-width: 768px) {
+    justify-content: center;
+    align-items: center;
   }
 `;
 
@@ -55,55 +36,6 @@ function AuthContainer({ children }: Props) {
   );
 }
 
-const Form = styled.form`
-  flex-direction: column;
-`;
-
-const Label = styled.label`
-  display: block;
-  font-family: Poppins;
-  font-size: 20px;
-  color: #5b5a5a;
-`;
-
-const Input = styled.input`
-  display: block;
-  box-sizing: border-box;
-  border: 1px solid #c4c4c4;
-  border-radius: 10px;
-  height: 33px;
-  padding-left: 10px;
-
-  font-family: Poppins;
-  font-size: 20px;
-  text-align: left;
-
-  margin-top: 11px;
-  margin-bottom: 22px;
-
-  width: 100%;
-`;
-
-const Submit = styled.input`
-  background: #5f8f3e;
-  color: white;
-  cursor: pointer;
-
-  font-family: Poppins;
-  font-size: 20px;
-  text-align: center;
-
-  border-radius: 10px;
-  border-color: #5f8f3e;
-  border-style: solid;
-
-  margin-top: 6px;
-  margin-bottom: 10px;
-  min-width: 100%;
-  height: 36px;
-  align-self: center;
-`;
-
 const ErrorMsg = styled.p`
   color: red;
   font-size: 15px;
@@ -112,14 +44,4 @@ const ErrorMsg = styled.p`
   margin-top: -20px;
 `;
 
-export {
-  StyledBack,
-  AuthHeader,
-  AuthContainer,
-  AuthContent,
-  Form,
-  Input,
-  Label,
-  Submit,
-  ErrorMsg,
-};
+export { AuthContainer, ErrorMsg };
