+++ frontend/src/components/navigation/header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, Link } from 'react-router-dom';
 import styled from 'styled-components';
 import { BiArrowBack } from 'react-icons/bi';
 import NavBar from './navBar';
@@ -9,9 +9,9 @@ const Navigation = styled.div`
   flex-direction: row;
   align-items: center;
   justify-content: center;
-  padding: 20px 0 10px 0;
+  padding: 30px 0 10px 0;
   @media screen and (min-width: 768px) {
-    padding: 30px 0 20px 0;
+    padding: 40px 0 20px 0;
   }
 `;
 
@@ -19,6 +19,7 @@ const BackArrow = styled(BiArrowBack)`
   color: black;
   position: absolute;
   left: 20px;
+  top: 40px;
   font-size: 25px;
   cursor: pointer;
   @media screen and (min-width: 768px) {
@@ -41,21 +42,31 @@ const StyledHeader = styled.header`
 type headerPropTypes = {
   headerText: string;
   navbar?: boolean; // optional prop, default defined below
+  back?: string;
 };
 
 /* note for future devs: pages that can be accessed from the navbar
  * will pass navbar as a prop, otherwise the back arrow will show
  */
-export default function Header({ headerText, navbar }: headerPropTypes) {
+export default function Header({ headerText, navbar, back }: headerPropTypes) {
   const navigate = useNavigate();
+
+  const previous = back ? (
+    <Link to={back}>
+      <BackArrow />
+    </Link>
+  ) : (
+    <BackArrow onClick={() => navigate(-1)} />
+  );
 
   return (
     <Navigation>
-      {navbar ? <NavBar /> : <BackArrow onClick={() => navigate(-1)} />}
+      {navbar ? <NavBar /> : previous}
       <StyledHeader>{headerText}</StyledHeader>
     </Navigation>
   );
 }
 Header.defaultProps = {
   navbar: false,
+  back: '',
 };
